Delete main.css
This was causing an issue with jekyll and github where the css for the project would only display the css correctly after being built using --watch command and making changes to a file. For fresh jekyll builds it would incorrectly use this .css and not build correctly. Removing this .css file seems to correct the issue and allows jekyll to compile css correctly. I hope this is clear. Thank you! :)